feat: add number of vaccination centres to calendar
This commit adds a new text string for each date to also display the
number of vaccination centres that have available booking slots. This
allows users to know if all available slots are for a specific
vaccination centre, or not, easily. Potentially helpful if they are
monitoring slots/locations for a specific day.

Signed-off-by: Ramon Quitales <me@rquitales.com>